+++ src/App.js
@@ -1,7 +1,11 @@
-import { Console, Random } from "@woowacourse/mission-utils";
+import { Console } from "@woowacourse/mission-utils";
 import Lotto from "./Lotto.js";
 import InputPriceValidation from "./InputPriceValidation.js";
 import InputBonusValidation from "./InputBonusValidation.js";
+import { printLotto } from "./MakeLotto.js";
+import { matchLotto } from "./MatchLotto.js";
+import { matchResult } from "./MatchResult.js";
+import { printResult } from "./PrintResult.js";
 
 class App {
   async play() {
@@ -14,6 +18,7 @@ class App {
   }
 }
 
+//UI 로직
 async function InputPrice() {
   while (true) {
     try {
@@ -61,77 +66,6 @@ async function InputBonus(lottoNumber) {
       Console.print(`${error.message}`);
     }
   }
-}
-
-// 랜덤 로또 번호 생성
-function makeLotto() {
-  const randomLotto = Random.pickUniqueNumbersInRange(1, 45, 6);
-  randomLotto.sort((a, b) => a - b);
-  return randomLotto;
-}
-
-// 랜덤 로또 번호 출력과 저장
-function printLotto(numLotto) {
-  Console.print(String(numLotto) + "개를 구매했습니다.")
-  const randomLottoArr = []
-  for (let i = 0; i < numLotto; i++) {
-    const lottoNumbers = makeLotto();
-    Console.print("[" + lottoNumbers.join(", ") + "]");
-    randomLottoArr.push(lottoNumbers);
-  }
-  Console.print("");
-  return randomLottoArr;
-}
-
-function matchLotto(randomLottoArr, lotto, bonus) {
-  const matchList = []
-  for (let i = 0; i < randomLottoArr.length; i++) {
-    matchList.push(matchLottoNumber(randomLottoArr[i], lotto, bonus))
-  }
-  return matchList;
-}
-
-
-function matchLottoNumber(randomLottoArr, lotto, bonus) {
-  const match = { lottoMatch: 0, bonusMatch: 0 }
-  for (let i = 0; i < 6; i++) {
-    if (randomLottoArr.includes(parseInt(lotto[i]))) {
-      match["lottoMatch"]++;
-    }
-  }
-  if (match["lottoMatch"] === 5) {
-    if (randomLottoArr.includes(parseInt(bonus))) {
-      match["bonusMatch"] = 1;
-    }
-  }
-  return match;
-}
-
-const getLottoMatch = (arr, el) => arr.filter(value => value["lottoMatch"] === el).length;
-const getBonusMatch = (arr, el) => arr.filter(value => value["bonusMatch"] === el).length;
-
-function matchResult(matchList) {
-  const matchFifth = getLottoMatch(matchList, 3);
-  const matchFourth = getLottoMatch(matchList, 4);
-  const matchSecond = getBonusMatch(matchList, 1);
-  const matchThird = getLottoMatch(matchList, 5) - matchSecond;
-  const matchFirst = getLottoMatch(matchList, 6);
-
-  return [matchFifth, matchFourth, matchThird, matchSecond, matchFirst];
-}
-
-function printResult(resultList, budget) {
-  Console.print("당첨 통계");
-  Console.print("---");
-  Console.print("3개 일치 (5,000원) - " + resultList[0] + "개");
-  Console.print("4개 일치 (50,000원) - " + resultList[1] + "개");
-  Console.print("5개 일치 (1,500,000원) - " + resultList[2] + "개");
-  Console.print("5개 일치, 보너스 볼 일치 (30,000,000원) - " + resultList[3] + "개");
-  Console.print("6개 일치 (2,000,000,000원) - " + resultList[4] + "개");
-
-  const profit = 5000 * resultList[0] + 50000 * resultList[1] + 1500000 * resultList[2] + 30000000 * resultList[3] + 2000000000 * resultList[4];
-  const profitRate = Math.round((profit / budget * 100) * 100) / 100;
-  Console.print("총 수익률은 " + profitRate + "%입니다.");
 }
 
 export default App;
